sonarcloud suggestions incorporated.
- Visualization was renamed to visualization
- generatePrime was renamed to generate_prime
- Other sonarcloud complaints were taken care of
